Revert the certs extra function and deduplicate some parts
